crashdump (tools v4.2.1)

A Tool for Handling Erlang Crash Dumps

This module contains functions for decrypting encrypted Erlang crash dumps, which can be enabled with the --enable-encrypted-crash-dumps build flag.

Functions

decrypt(EncryptedFile, PrivateKey)
(since OTP 29.0)

-spec decrypt(EncryptedFile, PrivateKey) -> {ok, erlang:iovec()} | {error, Reason}
                 when
                     EncryptedFile :: file:name(),
                     PrivateKey :: public_key:private_key(),
                     Reason :: dynamic().

Decrypts the Erlang crash dump at EncryptedFile using PrivateKey.

decrypt(EncryptedFile, PrivateKey, DecryptedFile)
(since OTP 29.0)

-spec decrypt(EncryptedFile, PrivateKey, DecryptedFile) -> ok | {error, Reason}
                 when
                     EncryptedFile :: file:name(),
                     PrivateKey :: public_key:private_key(),
                     DecryptedFile :: file:name(),
                     Reason :: dynamic().

Decrypts the Erlang crash dump at EncryptedFile using PrivateKey, writing the result to DecryptedFile.

pem_decrypt(EncryptedFile, KeyFile)
(since OTP 29.0)

-spec pem_decrypt(EncryptedFile, KeyFile) -> {ok, erlang:iovec()} | {error, Reason}
                     when EncryptedFile :: file:name(), KeyFile :: file:name(), Reason :: dynamic().

This is a convenience wrapper around crashdump:decrypt/2 that handles key extraction from a given PEM file.

pem_decrypt(EncryptedFile, KeyFile, DecryptedFile)
(since OTP 29.0)

-spec pem_decrypt(EncryptedFile, KeyFile, DecryptedFile) -> ok | {error, Reason}
                     when
                         EncryptedFile :: file:name(),
                         KeyFile :: file:name(),
                         DecryptedFile :: file:name(),
                         Reason :: dynamic().

This is a convenience wrapper around crashdump:decrypt/3 that handles key extraction from a given PEM file.