cover (tools v3.5)

The module cover provides a set of functions for coverage analysis of Erlang programs, counting how many times each executable line of code is executed when a program is run.
An executable line contains an Erlang expression such as a matching or a function call. A blank line or a line containing a comment, function head or pattern in a case- or receive statement is not executable.

Coverage analysis can be used to verify test cases, making sure all relevant code is covered, and may also be helpful when looking for bottlenecks in the code.

Before any analysis can take place, the involved modules must be Cover compiled. This means that some extra information is added to the module before it is compiled into a binary which then is loaded. The source file of the module is not affected and no .beam file is created.

Each time a function in a Cover compiled module is called, information about the call is added to an internal database of Cover. The coverage analysis is performed by examining the contents of the Cover database. The output Answer is determined by two parameters, Level and Analysis.

Level = module

Answer = {Module,Value}, where Module is the module name.

Level = function

Answer = [{Function,Value}], one tuple for each function in the module. A function is specified by its module name M, function name F and arity A as a tuple {M,F,A}.

Level = clause

Answer = [{Clause,Value}], one tuple for each clause in the module. A clause is specified by its module name M, function name F, arity A and position in the function definition C as a tuple {M,F,A,C}.

Level = line

Answer = [{Line,Value}], one tuple for each executable line in the module. A line is specified by its module name M and line number in the source file N as a tuple {M,N}.

Analysis = coverage

Value = {Cov,NotCov} where Cov is the number of executable lines in the module, function, clause or line that have been executed at least once and NotCov is the number of executable lines that have not been executed.

Analysis = calls

Value = Calls which is the number of times the module, function, or clause has been called. In the case of line level analysis, Calls is the number of times the line has been executed.

Distribution

Cover can be used in a distributed Erlang system. One of the nodes in the system must then be selected as the main node, and all Cover commands must be executed from this node. The error reason not_main_node is returned if an interface function is called on one of the remote nodes.

Use cover:start/1 and cover:stop/1 to add or remove nodes. The same Cover compiled code will be loaded on each node, and analysis will collect and sum up coverage data results from all nodes.

To only collect data from remote nodes without stopping cover on those nodes, use cover:flush/1

If the connection to a remote node goes down, the main node will mark it as lost. If the node comes back it will be added again. If the remote node was alive during the disconnected periode, cover data from before and during this periode will be included in the analysis.

SEE ALSO

code(3), compile(3)

Functions

analyse/0
(since OTP 18.0)

Specs

analyse() -> {result, analyse_ok(), analyse_fail()} | {error, not_main_node}.

Performs analysis of one or more Cover compiled modules, as specified by Analysis and Level (see above), by examining the contents of the internal database.

Analysis defaults to coverage and Level defaults to function.

If Modules is an atom (one module), the return will be OneResult, else the return will be {result, Ok, Fail}.

If Modules is not given, all modules that have data in the cover data table, are analysed. Note that this includes both cover compiled modules and imported modules.

If a given module is not Cover compiled, this is indicated by the error reason {not_cover_compiled, Module}.

analyse/1
(since OTP 18.0)

Specs

analyse(Analysis) -> {result, analyse_ok(), analyse_fail()} | {error, not_main_node}
           when Analysis :: analysis();
       (Level) -> {result, analyse_ok(), analyse_fail()} | {error, not_main_node}
           when Level :: level();
       (Modules) -> OneResult | {result, analyse_ok(), analyse_fail()} | {error, not_main_node}
           when Modules :: modules(), OneResult :: one_result().

analyse/2
(since OTP 18.0)

Specs

analyse(Analysis, Level) -> {result, analyse_ok(), analyse_fail()} | {error, not_main_node}
           when Analysis :: analysis(), Level :: level();
       (Modules, Analysis) ->
           OneResult | {result, analyse_ok(), analyse_fail()} | {error, not_main_node}
           when Analysis :: analysis(), Modules :: modules(), OneResult :: one_result();
       (Modules, Level) ->
           OneResult | {result, analyse_ok(), analyse_fail()} | {error, not_main_node}
           when Level :: level(), Modules :: modules(), OneResult :: one_result().

analyse/3
(since OTP 18.0)

Specs

analyse(Modules, Analysis, Level) ->
           OneResult | {result, analyse_ok(), analyse_fail()} | {error, not_main_node}
           when
               Analysis :: analysis(),
               Level :: level(),
               Modules :: modules(),
               OneResult :: one_result().

analyse_to_file/0
(since OTP 18.0)

Specs

analyse_to_file() -> {result, analyse_file_ok(), analyse_file_fail()} | {error, not_main_node}.

Makes copies of the source file for the given modules, where it for each executable line is specified how many times it has been executed.

The output file OutFile defaults to Module.COVER.out, or Module.COVER.html if the option html was used.

If Modules is an atom (one module), the return will be Answer, else the return will be a list, {result, Ok, Fail}.

If Modules is not given, all modules that have da ta in the cover data table, are analysed. Note that this includes both cover compiled modules and imported modules.

If a module is not Cover compiled, this is indicated by the error reason {not_cover_compiled, Module}.

If the source file and/or the output file cannot be opened using file:open/2, the function returns {error, {file, File, Reason}} where File is the file name and Reason is the error reason.

If a module was cover compiled from the .beam file, that is, using compile_beam/1 or compile_beam_directory/0,1 ,it is assumed that the source code can be found in the same directory as the .beam file, in ../src relative to that directory, or using the source path in Module:module_info(compile). When using the latter, two paths are examined: first the one constructed by joining ../src and the tail of the compiled path below a trailing src component, then the compiled path itself. If no source code is found, this is indicated by the error reason {no_source_code_found, Module}.

analyse_to_file/1
(since OTP 18.0)

Specs

analyse_to_file(Modules) ->
                   Answer |
                   {result, analyse_file_ok(), analyse_file_fail()} |
                   {error, not_main_node}
                   when Modules :: modules(), Answer :: analyse_answer();
               (Options) ->
                   {result, analyse_file_ok(), analyse_file_fail()} | {error, not_main_node}
                   when Options :: [analyse_option()].

analyse_to_file/2
(since OTP 18.0)

Specs

analyse_to_file(Modules, Options) ->
                   Answer |
                   {result, analyse_file_ok(), analyse_file_fail()} |
                   {error, not_main_node}
                   when
                       Modules :: modules(),
                       Options :: [analyse_option()],
                       Answer :: analyse_answer().

async_analyse_to_file/1
(since OTP R14B02)

Specs

async_analyse_to_file(Module) -> pid() when Module :: module().

This function works exactly the same way as analyse_to_file except that it is asynchronous instead of synchronous. The spawned process will link with the caller when created. If an error of type analyse_rsn() occurs while doing the cover analysis the process will crash with the same error reason as analyse_to_file would return.

async_analyse_to_file/2
(since OTP R14B02)

Specs

async_analyse_to_file(Module, OutFile) -> pid()
                         when Module :: module(), OutFile :: file:filename();
                     (Module, Options) -> pid()
                         when Module :: module(), Options :: [Option], Option :: html.

async_analyse_to_file/3
(since OTP R14B02)

Specs

async_analyse_to_file(Module, OutFile, Options) -> pid()
                         when
                             Module :: module(),
                             OutFile :: file:filename(),
                             Options :: [Option],
                             Option :: html.

compile/1

Specs

compile(ModFiles) -> Result | [Result] when ModFiles :: mod_files(), Result :: compile_result().

See compile:file/2.

Compiles a module for Cover analysis. The module is given by its module name Module or by its file name File. The .erl extension may be omitted. If the module is located in another directory, the path has to be specified.

Options is a list of compiler options which defaults to []. Only options defining include file directories and macros are passed to compile:file/2, everything else is ignored.

If the module is successfully Cover compiled, the function returns {ok, Module}. Otherwise the function returns {error, File}. Errors and warnings are printed as they occur.

If a list of ModFiles is given as input, a list of Result will be returned. The order of the returned list is undefined.

Note that the internal database is (re-)initiated during the compilation, meaning any previously collected coverage data for the module will be lost.

compile/2

Specs

compile(ModFiles, Options) -> Result | [Result]
           when ModFiles :: mod_files(), Options :: [option()], Result :: compile_result().

compile_beam/1

Specs

compile_beam(ModFiles) -> Result | [Result]
                when ModFiles :: beam_mod_files(), Result :: compile_beam_result().

Does the same as compile/1,2, but uses an existing .beam file as base, that is, the module is not compiled from source. Thus compile_beam/1 is faster than compile/1,2.

Note that the existing .beam file must contain abstract code, that is, it must have been compiled with the debug_info option. If not, the error reason {no_abstract_code, BeamFile} is returned. If the abstract code is encrypted, and no key is available for decrypting it, the error reason {encrypted_abstract_code, BeamFile} is returned.

If only the module name (that is, not the full name of the .beam file) is given to this function, the .beam file is found by calling code:which(Module). If no .beam file is found, the error reason non_existing is returned. If the module is already cover compiled with compile_beam/1, the .beam file will be picked from the same location as the first time it was compiled. If the module is already cover compiled with compile/1,2, there is no way to find the correct .beam file, so the error reason {already_cover_compiled, no_beam_found, Module} is returned.

{error, BeamFile} is returned if the compiled code cannot be loaded on the node.

If a list of ModFiles is given as input, a list of Result will be returned. The order of the returned list is undefined.

compile_beam_directory/0

Specs

compile_beam_directory() -> [Result] | {error, Reason}
                          when Reason :: file_error(), Result :: compile_beam_result().

Compiles all modules (.beam files) in a directory Dir for Cover analysis the same way as compile_beam/1 and returns a list with the return values.

Dir defaults to the current working directory.

The function returns {error, eacces} if the directory is not readable or {error, enoent} if the directory does not exist.

compile_beam_directory/1

Specs

compile_beam_directory(Dir) -> [Result] | {error, Reason}
                          when
                              Dir :: file:filename(),
                              Reason :: file_error(),
                              Result :: compile_beam_result().

compile_directory/0

Specs

compile_directory() -> [Result] | {error, Reason}
                     when Reason :: file_error(), Result :: compile_result().

See compile:file/2.

Compiles all modules (.erl files) in a directory Dir for Cover analysis the same way as compile_module/1,2 and returns a list with the return values.

Dir defaults to the current working directory.

The function returns {error, eacces} if the directory is not readable or {error, enoent} if the directory does not exist.

compile_directory/1

Specs

compile_directory(Dir) -> [Result] | {error, Reason}
                     when
                         Dir :: file:filename(),
                         Reason :: file_error(),
                         Result :: compile_result().

compile_directory/2

Specs

compile_directory(Dir, Options) -> [Result] | {error, Reason}
                     when
                         Dir :: file:filename(),
                         Options :: [option()],
                         Reason :: file_error(),
                         Result :: compile_result().

compile_module/1

Specs

compile_module(ModFiles) -> Result | [Result]
                  when ModFiles :: mod_files(), Result :: compile_result().

compile_module/2

Specs

compile_module(ModFiles, Options) -> Result | [Result]
                  when ModFiles :: mod_files(), Options :: [option()], Result :: compile_result().

export/1

Specs

export(File) -> ok | {error, Reason} when File :: file:filename(), Reason :: export_reason().

Exports the current coverage data for Module to the file ExportFile. It is recommended to name the ExportFile with the extension .coverdata, since other filenames cannot be read by the web based interface to cover.

If Module is not given, data for all Cover compiled or earlier imported modules is exported.

This function is useful if coverage data from different systems is to be merged.

See also import/1.

export/2

Specs

export(File, Module) -> ok | {error, Reason}
          when File :: file:filename(), Module :: module(), Reason :: export_reason().

flush/1
(since OTP R16B)

Specs

flush(Nodes) -> ok | {error, not_main_node} when Nodes :: node() | [node()].

Fetch data from the Cover database on the remote nodes and stored on the main node.

import/1

Specs

import(ExportFile) -> ok | {error, Reason}
          when
              ExportFile :: file:filename(),
              Reason :: {cant_open_file, ExportFile, FileReason :: term()} | not_main_node.

Imports coverage data from the file ExportFile created with export/1,2. Any analysis performed after this will include the imported data.

Note that when compiling a module all existing coverage data is removed, including imported data. If a module is already compiled when data is imported, the imported data is added to the existing coverage data.

Coverage data from several export files can be imported into one system. The coverage data is then added up when analysing.

Coverage data for a module cannot be imported from the same file twice unless the module is first reset or compiled. The check is based on the filename, so you can easily fool the system by renaming your export file.

See also export/1,2.

imported/0

Specs

imported() -> [file:filename()] | {error, not_main_node}.

Returns a list with all imported files.

imported_modules/0

Specs

imported_modules() -> [module()] | {error, not_main_node}.

Returns a list with all modules for which there are imported data.

is_compiled/1

Specs

is_compiled(Module) -> {file, File :: file:filename()} | false | {error, not_main_node}
               when Module :: module().

Returns {file, File} if the module Module is Cover compiled, or false otherwise. File is the .erl file used by compile_module/1,2 or the .beam file used by compile_beam/1.

local_only/0
(since OTP 22.0)

Specs

local_only() -> ok | {error, too_late}.

Only support running Cover on the local node. This function must be called before any modules have been compiled or any nodes added. When running in this mode, modules will be Cover compiled in a more efficient way, but the resulting code will only work on the same node they were compiled on.

modules/0

Specs

modules() -> [module()] | {error, not_main_node}.

Returns a list with all modules that are currently Cover compiled.

reset/0

Specs

reset() -> ok | {error, not_main_node}.

Resets all coverage data for a Cover compiled module Module in the Cover database on all nodes. If the argument is omitted, the coverage data will be reset for all modules known by Cover.

If Module is not Cover compiled, the function returns {error, {not_cover_compiled, Module}}.

reset/1

Specs

reset(Module) -> ok | {error, not_main_node} | {error, {not_cover_compiled, Module}}
         when Module :: module().

start/0

Specs

start() -> {ok, pid()} | {error, Reason} when Reason :: {already_started, pid()} | term().

Starts the Cover server which owns the Cover internal database. This function is called automatically by the other functions in the module.

start/1

Specs

start(Nodes) -> {ok, StartedNodes} | {error, not_main_node} | {error, local_only}
         when Nodes :: node() | [node()], StartedNodes :: [node()].

Starts a Cover server on the each of given nodes, and loads all cover compiled modules. This call will fail if cover:local_only/0 has been called.

stop/0

Specs

stop() -> ok | {error, not_main_node}.

Stops the Cover server and unloads all Cover compiled code.

stop/1

Specs

stop(Nodes) -> ok | {error, not_main_node} when Nodes :: node() | [node()].

Stops the Cover server and unloads all Cover compiled code on the given nodes. Data stored in the Cover database on the remote nodes is fetched and stored on the main node.

which_nodes/0

Specs

which_nodes() -> [node()].

Returns a list with all nodes that are part of the coverage analysis. Note that the current node is not returned. This node is always part of the analysis.