lcnt (tools v3.5)

The lcnt module is used to profile the internal ethread locks in the Erlang Runtime System. With lcnt enabled, internal counters in the runtime system are updated each time a lock is taken. The counters stores information about the number of acquisition tries and the number of collisions that has occurred during the acquisition tries. The counters also record the waiting time a lock has caused for a blocked thread when a collision has occurred.

The data produced by the lock counters will give an estimate on how well the runtime system will behave from a parallelizable view point for the scenarios tested. This tool was mainly developed to help Erlang runtime developers iron out potential and generic bottlenecks.

Locks in the emulator are named after what type of resource they protect and where in the emulator they are initialized, those are lock 'classes'. Most of those locks are also instantiated several times, and given unique identifiers, to increase locking granularity. Typically an instantiated lock protects a disjunct set of the resource, for example ets tables, processes or ports. In other cases it protects a specific range of a resource, for example pix_lock which protects index to process mappings, and is given a unique number within the class. A unique lock in lcnt is referenced by a name (class) and an identifier: {Name, Id}.

Some locks in the system are static and protects global resources, for example bif_timers and the run_queue locks. Other locks are dynamic and not necessarily long lived, for example process locks and ets-table locks. The statistics data from short lived locks can be stored separately when the locks are deleted. This behavior is by default turned off to save memory but can be turned on via lcnt:rt_opt({copy_save, true}). The lcnt:apply/1,2,3 functions enables this behavior during profiling.

See Also

LCNT User's Guide

Functions

apply/1
(since OTP R13B04)

Specs

apply(Fun) -> term() when Fun :: fun().

Same as apply(Fun, []).

apply/2
(since OTP R13B04)

Specs

apply(Fun, Args) -> term() when Fun :: fun(), Args :: [term()].

Clears the lock counters and then setups the instrumentation to save all destroyed locks. After setup the function is called, passing the elements in Args as arguments. When the function returns the statistics are immediately collected to the server. After the collection the instrumentation is returned to its previous behavior. The result of the applied function is returned.

This function should only be used for micro-benchmarks; it sets copy_save to true for the duration of the call, which can quickly lead to running out of memory.

apply/3
(since OTP R13B04)

Specs

apply(Module, Function, Args) -> term()
         when Module :: module(), Function :: atom(), Args :: [term()].

Same as apply(fun() -> erlang:apply(Module, Function, Args) end).

clear/0
(since OTP R13B04)

Specs

clear() -> ok.

Same as clear(node()).

clear/1
(since OTP R13B04)

Specs

clear(Node) -> ok when Node :: node().

Clears the internal lock statistics from the runtime system. This does not clear the data on the server only on runtime system. All counters for static locks are zeroed, all dynamic locks currently alive are zeroed and all saved locks now destroyed are removed. It also resets the duration timer.

collect/0
(since OTP R13B04)

Specs

collect() -> ok.

Same as collect(node()).

collect/1
(since OTP R13B04)

Specs

collect(Node) -> ok when Node :: node().

Collects lock statistics from the runtime system. The function starts a server if it is not already started. It then populates the server with lock statistics. If the server held any lock statistics data before the collect then that data is lost.

conflicts/0
(since OTP R13B04)

Specs

conflicts() -> ok.

Same as conflicts([]).

conflicts/1
(since OTP R13B04)

Specs

conflicts(Options) -> ok when Options :: [option()].

Prints a list of internal locks and its statistics.

For option description, see lcnt:inspect/2.

information/0
(since OTP R13B04)

Specs

information() -> ok.

Prints lcnt server state and generic information about collected lock statistics.

inspect/1
(since OTP R13B04)

Specs

inspect(Lock) -> ok
           when
               Lock :: Name | {Name, Id | [Id]},
               Name :: atom() | pid() | port(),
               Id :: atom() | integer() | pid() | port().

Same as inspect(Lock, []).

inspect/2
(since OTP R13B04)

Specs

inspect(Lock, Options) -> ok
           when
               Lock :: Name | {Name, Id | [Id]},
               Name :: atom() | pid() | port(),
               Id :: atom() | integer() | pid() | port(),
               Options :: [option()].

Prints a list of internal lock counters for a specific lock.

Lock Name and Id for ports and processes are interchangeable with the use of lcnt:swap_pid_keys/0 and is the reason why pid() and port() options can be used in both Name and Id space. Both pids and ports are special identifiers with stripped creation and can be recreated with lcnt:pid/2,3 and lcnt:port/1,2.

Option description:

{combine, boolean()}

Combine the statistics from different instances of a lock class.
Default: true

{locations, boolean()}

Print the statistics by source file and line numbers.
Default: false

{max_locks, MaxLocks}

Maximum number of locks printed or no limit with none.
Default: 20

{print, PrintOptions}

Printing options:

name

Named lock or named set of locks (classes). The same name used for initializing the lock in the VM.

id

Internal id for set of locks, not always unique. This could be table name for ets tables (db_tab), port id for ports, integer identifiers for allocators, etc.

type

Type of lock: rw_mutex, mutex, spinlock, rw_spinlock or proclock.

entry

In combination with {locations, true} this option prints the lock operations source file and line number entry-points along with statistics for each entry.

tries

Number of acquisitions of this lock.

colls

Number of collisions when a thread tried to acquire this lock. This is when a trylock is EBUSY, a write try on read held rw_lock, a try read on write held rw_lock, a thread tries to lock an already locked lock. (Internal states supervises this).

ratio

The ratio between the number of collisions and the number of tries (acquisitions) in percentage.

time

Accumulated waiting time for this lock. This could be greater than actual wall clock time, it is accumulated for all threads. Trylock conflicts does not accumulate time.

duration

Percentage of accumulated waiting time of wall clock time. This percentage can be higher than 100% since accumulated time is from all threads.

Default: [name,id,tries,colls,ratio,time,duration]

{reverse, boolean()}

Reverses the order of sorting.
Default: false

{sort, Sort}

Column sorting orders.
Default: time

{thresholds, Thresholds}

Filtering thresholds. Anything values above the threshold value are passed through.
Default: [{tries, 0}, {colls, 0}, {time, 0}]

load/1
(since OTP R13B04)

Specs

load(Filename) -> ok when Filename :: file:filename().

Restores previously saved data to the server.

locations/0
(since OTP R13B04)

Specs

locations() -> ok.

Same as locations([]).

locations/1
(since OTP R13B04)

Specs

locations(Options) -> ok when Options :: [option()].

Prints a list of internal lock counters by source code locations.

For option description, see lcnt:inspect/2.

pid/2
(since OTP R13B04)

Specs

pid(Id, Serial) -> pid() when Id :: integer(), Serial :: integer().

Same as pid(node(), Id, Serial).

pid/3
(since OTP R13B04)

Specs

pid(Node, Id, Serial) -> pid() when Node :: node(), Id :: integer(), Serial :: integer().

Creates a process id with creation 0.

port/1
(since OTP R13B04)

Specs

port(Id) -> port() when Id :: integer().

Same as port(node(), Id).

port/2
(since OTP R13B04)

Specs

port(Node, Id) -> port() when Node :: node(), Id :: integer().

Creates a port id with creation 0.

rt_clear/0
(since OTP R13B04)

Specs

rt_clear() -> ok.

Same as rt_clear(node()).

rt_clear/1
(since OTP R13B04)

Specs

rt_clear(Node) -> ok when Node :: node().

Clear the internal counters. Same as lcnt:clear(Node).

rt_collect/0
(since OTP R13B04)

Specs

rt_collect() -> [lock_counter_data()].

Same as rt_collect(node()).

rt_collect/1
(since OTP R13B04)

Specs

rt_collect(Node) -> [lock_counter_data()] when Node :: node().

Returns a list of raw lock counter data.

rt_mask/0
(since OTP 20.1)

Specs

rt_mask() -> [category_atom()].

Same as rt_mask(node()).

rt_mask/1
(since OTP 20.1)

Specs

rt_mask(Node) -> [category_atom()] when Node :: node();
       (Categories) -> ok | {error, copy_save_enabled} when Categories :: [category_atom()].

Refer to rt_mask/2. for a list of valid categories. All categories are enabled by default.

rt_mask/1
(since OTP 20.1)

Specs

rt_mask(Node) -> [category_atom()] when Node :: node();
       (Categories) -> ok | {error, copy_save_enabled} when Categories :: [category_atom()].

Same as rt_mask(node(), Categories).

rt_mask/2
(since OTP 20.1)

Specs

rt_mask(Node, Categories) -> ok | {error, copy_save_enabled}
           when Node :: node(), Categories :: [category_atom()].

Sets the lock category mask to the given categories.

This will fail if the copy_save option is enabled; see lcnt:rt_opt/2.

Valid categories are:

allocator

db (ETS tables)

debug

distribution

generic

io

process

scheduler

This list is subject to change at any time, as is the category any given lock may belong to.

rt_opt/1
(since OTP R13B04)

Specs

rt_opt(Option) -> boolean()
          when Option :: {Type, Value :: boolean()}, Type :: copy_save | process_locks.

Same as rt_opt(node(), {Type, Value}).

rt_opt/2
(since OTP R13B04)

Specs

rt_opt(Node, Option) -> boolean()
          when
              Node :: node(),
              Option :: {Type, Value :: boolean()},
              Type :: copy_save | process_locks.

Option description:

{copy_save, boolean()}

Retains the statistics of destroyed locks.
Default: false

This option will use a lot of memory when enabled, which must be reclaimed with lcnt:rt_clear. Note that it makes no distinction between locks that were destroyed and locks for which counting was disabled, so enabling this option will disable changes to the lock category mask.

{process_locks, boolean()}

Profile process locks, equal to adding process to the lock category mask; see lcnt:rt_mask/2
Default: true

save/1
(since OTP R13B04)

Specs

save(Filename) -> ok when Filename :: file:filename().

Saves the collected data to file.

start/0
(since OTP R13B04)

Specs

start() -> {ok, Pid} | {error, {already_started, Pid}} when Pid :: pid().

Starts the lock profiler server. The server only act as a medium for the user and performs filtering and printing of data collected by lcnt:collect/1.

stop/0
(since OTP R13B04)

Specs

stop() -> ok.

Stops the lock profiler server.

swap_pid_keys/0
(since OTP R13B04)

Specs

swap_pid_keys() -> ok.

Swaps places on Name and Id space for ports and processes.