tags (tools v4.0)

Generate Emacs TAGS file from Erlang source files

A TAGS file is used by Emacs to find function and variable definitions in any source file in large projects. This module can generate a TAGS file from Erlang source files. It recognises functions, records, and macro definitions.

OPTIONS

The functions above have an optional argument, Options. It is a list which can contain the following elements:

{outfile, NameOfTAGSFile} Create a TAGS file named NameOfTAGSFile.

{outdir, NameOfDirectory} Create a file named TAGS in the directory NameOfDirectory.

The default behaviour is to create a file named TAGS in the current directory.

Examples

tags:root([{outfile, "root.TAGS"}]).

This command will create a file named root.TAGS in the current directory. The file will contain references to all Erlang source files in the Erlang distribution.

tags:files(["foo.erl", "bar.erl", "baz.erl"], [{outdir, "../projectdir"}]).

Here we create file named TAGS placed it in the directory ../projectdir. The file contains information about the functions, records, and macro definitions of the three files.

SEE ALSO

Richard M. Stallman. GNU Emacs Manual, chapter "Editing Programs", section "Tag Tables". Free Software Foundation, 1995.

Anders Lindgren. The Erlang editing mode for Emacs. Ericsson, 1998.

Types

option()
(not exported)

-type option() ::
    {outfile, NameOfTAGSFile :: file:filename()} | {outdir, NameOfDirectory :: file:filename()}.

Functions

dir(Dir)

-spec dir(Dir) -> ok | error when Dir :: file:filename().

Equivalent to dir/2

dir(Dir, Options)

-spec dir(Dir, Options) -> ok | error when Dir :: file:filename(), Options :: [option()].

Create a TAGS file for all files in directory Dir.

dirs(DirList)

-spec dirs(DirList) -> ok | error when DirList :: [file:filename()].

Equivalent to dirs/2

dirs(DirList, Options)

-spec dirs(DirList, Options) -> ok | error when DirList :: [file:filename()], Options :: [option()].

Create a TAGS file for all files in any directory in DirList.

file(File)

-spec file(File) -> ok | error when File :: file:filename().

Equivalent to file/2

file(File, Options)

-spec file(File, Options) -> ok | error when File :: file:filename(), Options :: [option()].

Create a TAGS file for the file File.

files(FileList)

-spec files(FileList) -> ok | error when FileList :: [file:filename()].

Equivalent to files/2

files(FileList, Options)

-spec files(FileList, Options) -> ok | error when FileList :: [file:filename()], Options :: [option()].

Create a TAGS file for the files in the list FileList.

root()

-spec root() -> ok | error.

Equivalent to root/1

root(Options)

-spec root(Options) -> ok | error when Options :: [option()].

Create a TAGS file covering all files in the Erlang distribution.

subdir(Dir)

-spec subdir(Dir) -> ok | error when Dir :: file:filename().

Equivalent to subdir/2

subdir(Dir, Options)

-spec subdir(Dir, Options) -> ok | error when Dir :: file:filename(), Options :: [option()].

Descend recursively down the directory Dir and create a TAGS file based on all files found.

subdirs(DirList)

-spec subdirs(DirList) -> ok | error when DirList :: [file:filename()].

Equivalent to subdirs/2

subdirs(DirList, Options)

-spec subdirs(DirList, Options) -> ok | error when DirList :: [file:filename()], Options :: [option()].

Descend recursively down all the directories in DirList and create a TAGS file based on all files found.